+++ src/onegov.town/onegov/town/assets/js/editor.js
@@ -1,3 +1,9 @@
 $(function() {
-    $('textarea.editor').redactor();
+    _.each($('textarea.editor'), function(el) {
+        var textarea = $(el);
+
+        textarea.redactor({
+            imageUpload: textarea.closest('form').data('image-upload-url')
+        });
+    });
 });