+++ src/onegov.town/onegov/town/views/images.py
@@ -34,9 +34,11 @@ def view_get_image_collection(self, request):
     }
 
 
-@TownApp.view(model=ImageCollection, name='upload', request_method='POST',
-              permission=Private)
-def view_upload_file(self, request):
+def handle_file_upload(self, request):
+    """ Stores the file given with the request and returns the url to the
+    resulting file.
+
+    """
     extension = request.params['file'].filename.split('.')[-1]
 
     if extension not in {'png', 'jpg', 'jpeg', 'gif', 'svg'}:
@@ -46,4 +48,22 @@ def view_upload_file(self, request):
 
     self.filestorage.setcontents(filename, request.params['file'].file.read())
 
+    return request.filestorage_link(self.path_prefix + filename)
+
+
+@TownApp.view(model=ImageCollection, name='upload', request_method='POST',
+              permission=Private)
+def view_upload_file(self, request):
+    request.assert_valid_csrf_token()
+
+    handle_file_upload(self, request)
+
     return morepath.redirect(request.link(self))
+
+
+@TownApp.json(model=ImageCollection, name='upload.json', request_method='POST',
+              permission=Private)
+def view_upload_file_by_json(self, request):
+    request.assert_valid_csrf_token()
+
+    return {'filelink': handle_file_upload(self, request)}